+++ src/form/range.hpp
@@ -35,4 +35,5 @@ class RangeMap : public std::map<int64_t, Range> {
   void prune();
 
   std::string toString() const;
+  std::string toString(int64_t index) const;
 };